test(ambrosian): ordo-validation of temporale anchors (Year C 2024-25, Year A 2025-26)
Add AmbrosianTemporaleOrdoValidationTest (@group slow), pinning the
AmbrosianTemporale engine's anchor-block output across civil years
2024, 2025, and 2026, which compose the two liturgical years the
chiesadimilano.it 2024-edition daily-liturgy widget covers. Extract
the buildContext()/runEngine() harness shared with AmbrosianTemporaleTest
into AmbrosianTemporaleHarnessTrait to avoid duplicating it.

Spot-checked Advent I (both years), Dedication of the Duomo, and
Christ the King against chiesadimilano.it: all four exact-date checks
matched with no discrepancy; an after-Pentecost bound check also
brackets correctly. Findings recorded in the Plan 3 doc under a new
"Ordo-validation findings" section.
