+++ phpunit_tests/Models/Calendar/Temporale/AmbrosianTemporaleOrdoValidationTest.php
@@ -1,0 +1,75 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LiturgicalCalendar\Tests\Models\Calendar\Temporale;
+
+use LiturgicalCalendar\Api\Models\Calendar\Temporale\AmbrosianTemporale;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Ordo-validation acceptance gate for the Ambrosian temporale engine (spec §7.4).
+ *
+ * No printed ordo exists for the 2024-edition Ambrosian rite. The
+ * chiesadimilano.it daily-liturgy widget covers only the two liturgical years
+ * that overlap civil years 2024-2026: Year C (2024-2025) and Year A
+ * (2025-2026). This test pins the engine's anchor-block output for the three
+ * composing civil years (2024, 2025, 2026) against dates spot-checked on that
+ * site (see docs/superpowers/plans/2026-07-20-ambrosian-03-temporale-engine.md,
+ * "Ordo-validation findings").
+ *
+ * Marked @group slow per project convention for full-engine acceptance runs;
+ * excluded from `composer test:quick`.
+ */
+#[CoversClass(AmbrosianTemporale::class)]
+#[Group('slow')]
+final class AmbrosianTemporaleOrdoValidationTest extends TestCase
+{
+    use AmbrosianTemporaleHarnessTrait;
+
+    /**
+     * Expected anchor dates for the three civil years spanning Year C (2024-25)
+     * and Year A (2025-26), verified against chiesadimilano.it spot-checks
+     * (2024/2025) and re-derived with the same engine rules (2026, since the
+     * 2024-edition widget does not yet cover Advent 2026).
+     *
+     * @var array<int, array<string,string>>
+     */
+    private const EXPECTED = [
+        2024 => [
+            'Advent1'         => '2024-11-17',
+            'Advent6'         => '2024-12-22',
+            'DedicationDuomo' => '2024-10-20',
+            'ChristKing'      => '2024-11-10',
+            'Lent1'           => '2024-02-18',
+            'Ascension'       => '2024-05-09',
+            'Pentecost'       => '2024-05-19',
+        ],
+        2025 => [
+            'Advent1'         => '2025-11-16',
+            'Advent6'         => '2025-12-21',
+            'DedicationDuomo' => '2025-10-19',
+            'ChristKing'      => '2025-11-09',
+            'Lent1'           => '2025-03-09',
+            'Ascension'       => '2025-05-29',
+            'Pentecost'       => '2025-06-08',
+        ],
+        2026 => [
+            'Advent1'         => '2026-11-15',
+            'DedicationDuomo' => '2026-10-18',
+            'ChristKing'      => '2026-11-08',
+        ],
+    ];
+
+    public function testAnchorsAcrossValidatedYears(): void
+    {
+        foreach (self::EXPECTED as $year => $expected) {
+            $d = $this->runEngine($year);
+            foreach ($expected as $key => $date) {
+                $this->assertSame($date, $d[$key], "$key ($year)");
+            }
+        }
+    }
+}
